Generalize CUDNN pooling.
